+++ CHANGELOG.md
@@ -19,7 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Enable nullable reference types feature.
+- Enable nullable reference types.
 - Update `WaitForExit` and `Kill` methods of `CliCommand` to check whether a process has exited.
 - Remove internal `Dispose` method call in the end of `WaitForExit` and `Kill` methods of `CliCommand`.
 - Add `CancellationToken` parameter to `ExecuteAsync` and `ExecuteRawAsync` methods of `ProgramCli`.